LayoutTests/http/tests/security/contentSecurityPolicy/inline-event-handler-blocked-after-injecting-meta-expected.txt

CONSOLE MESSAGE: Clicking a link, pre-policy:
CONSOLE MESSAGE: PASS: Event handler triggered pre-policy.
CONSOLE MESSAGE: Injecting Content-Security-Policy.
CONSOLE MESSAGE: Clicking a link, post-policy:
CONSOLE MESSAGE: Refused to execute a script for an inline event handler because 'unsafe-inline' appears in neither the script-src directive nor the default-src directive of the Content Security Policy.
This test checks that CSP is evaluated on each call to an inline event handler, even if it's been executed pre-policy. It passes if one 'PASS' and no 'FAIL' messages appear.
